// apps/web/src/types/news.types.ts
export interface Article {
  field: string;
  region: string;
  date: number;
  id: string;
  title: string;
  slug: string;
  featured: boolean;
  summary: string;
  imageUrl: string;
  author: string;
  source: string;
  readTime: number;
  externalUrl: string;
  sourceType: string;
}

export type State = {
  field: string;
  region: string;
  time: string;
};